// v1/src/types/Payables.ts
export type Payable = z.infer<typeof payableSchema>

export type IPayables = Payable &
  abstractType & {
    id: number
    supplier: ISupplier
    collaborator: ICollaborator
    payableStatus: PayableStatus
    totalValue: number
    approvers: number[]
    recurenceData:
      | {
          recurrenceType: RecurenceType | undefined
          startDate: Date | undefined
          endDate: Date | undefined
          dueDay: number | undefined
        }
      | undefined
    files: {
      id: number
      fileUrl: string
    }[]
    approvals: {
      id: number
      approved: boolean
      obs: string
      collaborator: Pick<ICollaborator, 'id' | 'name' | 'email'>
      user: Pick<IUser, 'id' | 'email' | 'name'>
    }[]
    categorization: {
      program: Pick<IProgram, 'id' | 'name'>
      budgetPlan: Pick<BudgetPlan, 'id' | 'year' | 'scenarioName' | 'version'>
      costCenter: Pick<ICostCenter, 'id' | 'name' | 'active'>
      costCenterCategory: Pick<ICostCenterCategory, 'id' | 'name' | 'active'>
      costCenterSubCategory: Pick<ICostCenterSubCategory, 'id' | 'name' | 'type'>
    }
    installments: Installments[]
    contract: IContract
    createdAt: string
    updatedAt: string
  }